app-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[dotenv]
python-dotenv